+++ stories/uikit/biz/Typography.stories.tsx
@@ -1,0 +1,57 @@
+import type { Meta, StoryObj, StoryFn } from '@storybook/react'
+import { Stack } from '@tidbcloud/uikit'
+import { Typography, TYPOGRAPHY_STYLES_MAP } from '@tidbcloud/uikit/biz'
+
+import { COLOR_LIST } from '../../constants'
+
+type Story = StoryObj<typeof Typography>
+
+const decorator = (Story: StoryFn) => {
+  return (
+    <div style={{ margin: '3em' }}>
+      <Story />
+    </div>
+  )
+}
+
+const meta: Meta<typeof Typography> = {
+  title: 'Biz/Typography',
+  component: Typography,
+  decorators: [decorator],
+  tags: ['autodocs'],
+  parameters: {}
+}
+
+export default meta
+
+// More on interaction testing: https://storybook.js.org/docs/react/writing-tests/interaction-testing
+export const Primary: Story = {
+  parameters: {
+    controls: { expanded: true }
+  },
+  args: {
+    variant: 'label-lg',
+    children: 'This is a label-lg'
+  },
+  argTypes: {
+    c: {
+      options: ['', ...COLOR_LIST],
+      control: { type: 'select' }
+    },
+    variant: {
+      options: Object.keys(TYPOGRAPHY_STYLES_MAP),
+      control: { type: 'select' }
+    }
+  }
+}
+
+export const Variants: Story = {
+  render: () => {
+    const texts = Object.keys(TYPOGRAPHY_STYLES_MAP).map((i) => (
+      <Typography variant={i as any} key={i}>
+        this is variant: {i}
+      </Typography>
+    ))
+    return <Stack>{texts}</Stack>
+  }
+}